OrderedDict is part of Python 2.7
