+++ sabnzbd/rating.py
@@ -26,12 +26,22 @@ import time
 import logging
 import copy
 import socket
+import Queue
+import collections
 from threading import RLock, Thread
-
 import sabnzbd
 from sabnzbd.decorators import synchronized
-from sabnzbd.utils.ordered import OrderedSetQueue
 import sabnzbd.cfg as cfg
+
+# A queue which ignores duplicates but maintains ordering
+class OrderedSetQueue(Queue.Queue):
+    def _init(self, maxsize):
+        self.maxsize = maxsize
+        self.queue = collections.OrderedDict()
+    def _put(self, item):
+        self.queue[item] = None
+    def _get(self):
+        return self.queue.popitem()[0]
 
 _RATING_URL = "/releaseRatings/releaseRatings.php"
 RATING_LOCK = RLock()
